[ET-VK][op_tests][BE] Refactor codegen components to prepare for benchmark generation
## Context

Refactor operator test code generation scripts, such that components can be re-used to generate operator benchmarks.

In broad strokes, the refactors implemented by this diff are as follows:

* Improve granularity of Python modules
* Replace `test` with `correctness_test`, to make it clear that we are generating correctness tests. **Note that I haven't changed the top level target name `compute_graph_op_tests_bin` since I believe it would be too verbose.

Differential Revision: [D63286131](https://our.internmc.facebook.com/intern/diff/D63286131/)

[ghstack-poisoned]
